Stop reloading when in non-graphical target
Previous version would automatically reload to default (usually graphical) target when booting or manually switching. This caused problems when trying to boot to a non-default target such as a multi-user text console. This is updated so that the user is informed switching will not be done when in the non-default (usually text console or some such) target. Open to ideas on if/how this could be done better.