Change user prompt to adhere to prompt conventions
I believe the letter 'Y' should not be capitalized since that suggests
it being the default (just pressed Enter) action of the prompt.

Multiple tools in the UNIX world (apt, pacman) capitalize the default
action in these prompts. Then there are tools like ssh which do not
capitalize any options and don't have a default. This commit makes this
tool's prompts look similar to those of the ssh utility while not
suggesting there is a default option.

Closes #7