Enhance GPU removal
Added code to remove the PCIe bus IDs of the GPU after switching to the internal mode and reloading. Combined with unloading and reloading the driver, this allows both Nvidia and AMD GPUs to be physically unplugged after the display manager reloads. Previously this would result in a system hang. Tested on kernel 5.4.