Check if drivers are in use
Adds a check for if the drivers are still in use before unloading them. Previously an application like Folding @ Home could cause the reload to stop at a black screen. Now if the drivers are still in use after the display manager stops, then the display manager is simply reloaded and an error message is given without removing the drivers or egpu.